implemented pagination on championship resource

+++ src/main/kotlin/br/com/mirantebackend/dto/pageable/PageDto.kt
@@ -1,7 +1,7 @@
 package br.com.mirantebackend.dto.pageable
 
 class PageDto<T : RecordDto>(
-    var pageSize: Long,
+    var pageSize: Int,
     var pageNumber: Int,
     var total: Long,
     var records: Collection<T>?
